Add analyzer for missing types in InlineData.
When using InlineData annotations in Xunit, it is dangerous not to
annotate the parameters supplied by the annotations with their intended
types.
This is because Xunit only supports .NET primitives (and a few other
derived types), and will silently skip test discovery of any test that
uses InlineData with any other type.
This can cause tests to unexpectedly and silently stop running.

This commit adds an analyzer which warns about missing type annotations
on parameters in tests which use InlineData annotations.